feat: finish gradual rollout topic, refactor topics
